requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
PyJWT==2.8.0
PyJWT==2.8.0
pymongo==4.7.2
python-dotenv==1.0.1
pytz==2024.1
razorpay==1.4.2
twilio==9.2.0
